Retry all requests that fail with a 50x error
These errors are usually transient, and retrying might succeed.

We retry 4 times (for a maximum of 5 requests in total), using an
exponential backoff strategy. The maximum possible delay is thus 15
seconds.

To deactivate the retry mechanism or change the maximum number of
retries, adjust `api.NUM_RETRIES` after the initial import of the
library.