Switch from oauth2client to google-auth
This changes the OAuth2 functionality to use `google-auth` instead of
`oauth2client`, which is [deprecated][0].

It adds a new argument `credentials` to the `API` constructor, which
allows to either use an `Http` object as previously, or to directly
supply a `google.oauth2.credentials.Credentials` instance. Existing code
should thus continue to work unaffected.


[0]: https://google-auth.readthedocs.io/en/latest/oauth2client-deprecation.html